Add common access methods to dictionary repository
